seperate concerns for show/movie controllers
